+++ Cargo.toml
@@ -1,6 +1,6 @@
 [workspace]
 members = [
-    "dataprep",
+    "tomb",
     "dataprep-lib"
 ]
 [workspace.package]
