+++ tomb/src/main.rs
@@ -50,5 +50,6 @@ async fn main() {
         } => {
             unpack_pipeline(&input_dir, &output_dir).await.unwrap();
         }
+        _ => unimplemented!("todo...")
     }
 }
